Move commons to base to prevent partial initialization
